make searchbar into functional component and get searched records
