+++ src/components/navbar/Navbar.tsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 
 import UserMenu from "./UserMenu";
 import Dropdown from "./Dropdown";
-import SearchBar from "./SearchBar";
+import SearchBar from "../partials/SearchBar";
 
 export default class Navbar extends Component {
     render() {
